Refactor ContestCard component location and import paths
- Moved ContestCard component from ContestsList to the main components directory
- Updated import paths in related files
- Removed redundant ContestCountdown and ContestCard files from ContestsList
