+++ packages/commonwealth/client/scripts/views/components/ContestCard/ContestAlert/ContestAlert.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../styles/shared';
+@import '../../../../styles/shared';
 
 .ContestAlert {
   display: flex;
