+++ packages/commonwealth/client/scripts/views/components/ContestCard/ContestCard.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../styles/shared';
+@import '../../../styles/shared';
 
 .ContestCard {
   max-width: 600px;
