--- packages/commonwealth/client/scripts/views/pages/CommunityManagement/Contests/ContestsList/ContestCard/ContestCard.scss
@@ -1,168 +1,0 @@
-@import '../../../../../../styles/shared';
-
-.ContestCard {
-  max-width: 600px;
-  width: unset;
-  padding: 0;
-  position: relative;
-
-  &.isHorizontal {
-    flex-direction: row;
-    max-width: unset;
-
-    .contest-image {
-      border-radius: 5px 0 0 5px;
-      max-height: 240px;
-      max-width: 522px;
-    }
-  }
-
-  .active-contest-tag {
-    position: absolute;
-    top: 10px;
-    left: 10px;
-  }
-
-  .contest-image {
-    border-top-left-radius: 5px;
-    border-top-right-radius: 5px;
-    width: 100%;
-    max-height: 275px;
-    object-fit: cover;
-
-    @include extraSmall {
-      min-height: 195px;
-    }
-  }
-
-  .contest-body {
-    width: 100%;
-    padding: 16px;
-
-    .header-row {
-      display: flex;
-      justify-content: space-between;
-      align-items: start;
-      gap: 32px;
-
-      & > .Text {
-        word-break: break-word;
-      }
-
-      @include extraSmall {
-        flex-direction: column;
-        gap: 4px;
-      }
-    }
-
-    .topics {
-      color: $neutral-500;
-    }
-
-    .prizes-header {
-      margin-top: 8px;
-    }
-
-    .prizes {
-      display: flex;
-      gap: 8px;
-      flex-direction: column;
-
-      .prize-row {
-        display: flex;
-        justify-content: space-between;
-        gap: 8px;
-
-        .label {
-          color: $neutral-600;
-        }
-
-        .bold {
-          word-break: break-word;
-        }
-      }
-    }
-
-    .actions {
-      display: flex;
-      gap: 16px;
-      margin-top: 8px;
-
-      @include extraSmall {
-        justify-content: space-between;
-        margin-top: 12px;
-
-        .ThreadAction {
-          .Text.caption {
-            display: none;
-          }
-
-          svg {
-            width: 24px;
-            height: 24px;
-          }
-        }
-      }
-    }
-
-    .farcaster-cta {
-      width: 100%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 7px;
-      padding-block: 8px;
-      border: 1px solid $neutral-200;
-      border-radius: 6px;
-      margin-top: 8px;
-
-      &:not(:disabled):hover {
-        cursor: pointer;
-        background-color: $neutral-25;
-      }
-
-      &:not(:disabled):active {
-        cursor: pointer;
-        background-color: $neutral-50;
-      }
-
-      &:focus {
-        outline: none;
-      }
-
-      &:disabled {
-        cursor: not-allowed;
-        filter: grayscale(1);
-      }
-
-      img {
-        width: 32px;
-        height: 32px;
-      }
-    }
-  }
-
-  .contest-footer {
-    padding-inline: 16px;
-    padding-bottom: 16px;
-    width: 100%;
-
-    .Divider {
-      margin-bottom: 16px;
-    }
-
-    .buttons {
-      display: flex;
-      justify-content: flex-end;
-
-      @include extraSmall {
-        flex-direction: column;
-
-        .cta-btn,
-        .Button {
-          width: 100%;
-        }
-      }
-    }
-  }
-}
