--- packages/commonwealth/client/scripts/views/pages/CommunityManagement/Contests/ContestsList/ContestCard/index.ts
@@ -1,3 +1,0 @@
-import ContestCard from './ContestCard';
-
-export default ContestCard;
