--- packages/commonwealth/client/scripts/views/pages/CommunityManagement/Contests/ContestsList/ContestCountdown/index.ts
@@ -1,3 +1,0 @@
-import ContestCountdown from './ContestCountdown';
-
-export default ContestCountdown;
